+++ dns_dnssec.go
@@ -61,14 +61,14 @@ func (t *DNSSec) UnmarshalJSON(data []byte) error {
 }
 
 // AddAgent - Add agent to DNSSec test
-func (t *DNSSec) AddAgent(id int) {
-	agent := Agent{AgentID: Int(id)}
+func (t *DNSSec) AddAgent(id int64) {
+	agent := Agent{AgentID: Int64(id)}
 	*t.Agents = append(*t.Agents, agent)
 }
 
 // AddAlertRule - Adds an alert to agent test
-func (t *DNSSec) AddAlertRule(id int) {
-	alertRule := AlertRule{RuleID: Int(id)}
+func (t *DNSSec) AddAlertRule(id int64) {
+	alertRule := AlertRule{RuleID: Int64(id)}
 	*t.AlertRules = append(*t.AlertRules, alertRule)
 }
