+++ dns_trace.go
@@ -62,14 +62,14 @@ func (t *DNSTrace) UnmarshalJSON(data []byte) error {
 }
 
 // AddAgent - Add agent to DNS Trace test
-func (t *DNSTrace) AddAgent(id int) {
-	agent := Agent{AgentID: Int(id)}
+func (t *DNSTrace) AddAgent(id int64) {
+	agent := Agent{AgentID: Int64(id)}
 	*t.Agents = append(*t.Agents, agent)
 }
 
 // AddAlertRule - Adds an alert to agent test
-func (t *DNSTrace) AddAlertRule(id int) {
-	alertRule := AlertRule{RuleID: Int(id)}
+func (t *DNSTrace) AddAlertRule(id int64) {
+	alertRule := AlertRule{RuleID: Int64(id)}
 	*t.AlertRules = append(*t.AlertRules, alertRule)
 }
